Modularity; bumped crate versions
