+++ aleph-client/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "aleph_client"
-version = "1.9.0"
+version = "1.9.1"
 edition = "2021"
 license = "Apache 2.0"
