+++ e2e-tests/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "aleph-e2e-client"
-version = "0.4.0"
+version = "0.4.1"
 edition = "2021"
 license = "Apache 2.0"
