+++ e2e-tests/src/test/kick_out.rs
@@ -1,35 +1,29 @@
 use aleph_client::{
-    change_validators, get_committee_kick_out_config, get_current_era_validators,
-    get_current_session, get_kick_out_reason_for_validator, get_next_era_non_reserved_validators,
-    get_next_era_reserved_validators, get_next_era_validators,
-    get_underperformed_validator_session_count, kick_out_from_committee, wait_for_at_least_session,
-    wait_for_event, wait_for_full_era_completion, AccountId, AnyConnection, RootConnection,
-    SignedConnection, XtStatus,
+    get_current_era_validators, get_current_session, get_next_era_validators,
+    wait_for_at_least_session, SignedConnection,
 };
-use codec::Decode;
 use log::info;
 use primitives::{
-    BoundedVec, CommitteeKickOutConfig, CommitteeSeats, EraValidators, KickOutReason, SessionCount,
-    DEFAULT_CLEAN_SESSION_COUNTER_DELAY, DEFAULT_KICK_OUT_MINIMAL_EXPECTED_PERFORMANCE,
-    DEFAULT_KICK_OUT_SESSION_COUNT_THRESHOLD,
+    KickOutReason, SessionCount, DEFAULT_CLEAN_SESSION_COUNTER_DELAY,
+    DEFAULT_KICK_OUT_MINIMAL_EXPECTED_PERFORMANCE, DEFAULT_KICK_OUT_SESSION_COUNT_THRESHOLD,
 };
-use sp_runtime::Perbill;
 
 use crate::{
-    accounts::{account_ids_from_keys, get_validator_seed, NodeKeys},
+    accounts::{get_validator_seed, NodeKeys},
+    kick_out::{
+        check_committee_kick_out_config, check_kick_out_event,
+        check_underperformed_validator_reason, check_underperformed_validator_session_count,
+        check_validators, setup_test,
+    },
     rewards::set_invalid_keys_for_validator,
-    validators::get_test_validators,
     Config,
 };
 
 const VALIDATOR_TO_DISABLE_NON_RESERVED_INDEX: u32 = 0;
 const VALIDATOR_TO_DISABLE_OVERALL_INDEX: u32 = 2;
+// Address for //2 (Node2). Depends on the infrastructure setup.
 const NODE_TO_DISABLE_ADDRESS: &str = "127.0.0.1:9945";
 const SESSIONS_TO_MEET_KICK_OUT_THRESHOLD: SessionCount = 4;
-
-const VALIDATOR_TO_MANUALLY_KICK_OUT_NON_RESERVED_INDEX: u32 = 1;
-const VALIDATOR_TO_MANUALLY_KICK_OUT_OVERALL_INDEX: u32 = 3;
-const MANUAL_KICK_OUT_REASON: &str = "Manual kick out reason";
 
 /// Runs a chain, sets up a committee and validators. Sets an incorrect key for one of the
 /// validators. Waits for the offending validator to hit the kick-out threshold of sessions without
@@ -57,8 +51,8 @@ pub fn kick_out_automatic(config: &Config) -> anyhow::Result<()> {
         &non_reserved_validators[VALIDATOR_TO_DISABLE_NON_RESERVED_INDEX as usize];
     info!(target: "aleph-client", "Validator to disable: {}", validator_to_disable);
 
-    check_underperformed_validator_session_count(&root_connection, validator_to_disable, 0);
-    check_underperformed_validator_reason(&root_connection, validator_to_disable, &None);
+    check_underperformed_validator_session_count(&root_connection, validator_to_disable, &0);
+    check_underperformed_validator_reason(&root_connection, validator_to_disable, None);
 
     let validator_seed = get_validator_seed(VALIDATOR_TO_DISABLE_OVERALL_INDEX);
     let stash_controller = NodeKeys::from(validator_seed);
@@ -77,16 +71,21 @@ pub fn kick_out_automatic(config: &Config) -> anyhow::Result<()> {
         current_session + SESSIONS_TO_MEET_KICK_OUT_THRESHOLD,
     )?;
 
-    check_underperformed_validator_session_count(&root_connection, validator_to_disable, 0);
+    check_underperformed_validator_session_count(&root_connection, validator_to_disable, &0);
     let expected_kick_out_reason =
         KickOutReason::InsufficientUptime(DEFAULT_KICK_OUT_SESSION_COUNT_THRESHOLD);
     check_underperformed_validator_reason(
         &root_connection,
         validator_to_disable,
-        &Some(expected_kick_out_reason),
+        Some(&expected_kick_out_reason),
     );
 
-    let expected_non_reserved = &non_reserved_validators[(VALIDATOR_TO_DISABLE_NON_RESERVED_INDEX + 1) as usize..];
+    let expected_non_reserved =
+        &non_reserved_validators[(VALIDATOR_TO_DISABLE_NON_RESERVED_INDEX + 1) as usize..];
+
+    let expected_kicked_out_validators =
+        vec![(validator_to_disable.clone(), expected_kick_out_reason)];
+    check_kick_out_event(&root_connection, &expected_kicked_out_validators)?;
 
     // Check next era validators.
     check_validators(
@@ -96,8 +95,6 @@ pub fn kick_out_automatic(config: &Config) -> anyhow::Result<()> {
         get_next_era_validators,
     );
 
-    wait_for_kick_out_event(&root_connection)?;
-
     // Check current validators.
     check_validators(
         &root_connection,
@@ -106,199 +103,7 @@ pub fn kick_out_automatic(config: &Config) -> anyhow::Result<()> {
         get_current_era_validators,
     );
 
-    check_underperformed_validator_reason(&root_connection, validator_to_disable, &None);
+    check_underperformed_validator_reason(&root_connection, validator_to_disable, None);
 
     Ok(())
-}
-
-pub fn kick_out_manual(config: &Config) -> anyhow::Result<()> {
-    let (root_connection, reserved_validators, non_reserved_validators) = setup_test(config)?;
-
-    // Check current era validators.
-    let start_era_validators = check_validators(
-        &root_connection,
-        &reserved_validators,
-        &non_reserved_validators,
-        get_current_era_validators,
-    );
-
-    check_committee_kick_out_config(
-        &root_connection,
-        DEFAULT_KICK_OUT_MINIMAL_EXPECTED_PERFORMANCE,
-        DEFAULT_KICK_OUT_SESSION_COUNT_THRESHOLD,
-        DEFAULT_CLEAN_SESSION_COUNTER_DELAY,
-    );
-
-    let validator_to_manually_kick_out =
-        &non_reserved_validators[VALIDATOR_TO_MANUALLY_KICK_OUT_NON_RESERVED_INDEX as usize];
-    info!(target: "aleph-client", "Validator to manually kick out: {}", validator_to_manually_kick_out);
-
-    check_underperformed_validator_session_count(&root_connection, validator_to_manually_kick_out, 0);
-    check_underperformed_validator_reason(&root_connection, validator_to_manually_kick_out, &None);
-
-    let manual_reason = MANUAL_KICK_OUT_REASON.as_bytes().to_vec();
-    let reason = BoundedVec::try_from(manual_reason).expect("Incorrect manual kick out reason!");
-    let manual_kick_out_reason = KickOutReason::OtherReason(reason);
-
-    kick_out_from_committee(
-        &root_connection,
-        validator_to_manually_kick_out,
-        manual_kick_out_reason,
-        XtStatus::InBlock,
-    );
-
-    let current_session = get_current_session(&root_connection);
-
-    wait_for_at_least_session(
-        &root_connection,
-        current_session + SESSIONS_TO_MEET_KICK_OUT_THRESHOLD,
-    )?;
-
-    check_underperformed_validator_session_count(&root_connection, validator_to_manually_kick_out, 0);
-    check_underperformed_validator_reason(
-        &root_connection,
-        validator_to_manually_kick_out,
-        &Some(manual_kick_out_reason),
-    );
-
-    let expected_non_reserved = &non_reserved_validators[..VALIDATOR_TO_MANUALLY_KICK_OUT_NON_RESERVED_INDEX as usize]
-        .iter()
-        .chain(&non_reserved_validators[(VALIDATOR_TO_MANUALLY_KICK_OUT_NON_RESERVED_INDEX + 1) as usize..].iter())
-        .collect();
-
-    // Check next era validators.
-    check_validators(
-        &root_connection,
-        &reserved_validators,
-        &non_reserved_validators[(VALIDATOR_TO_DISABLE_NON_RESERVED_INDEX + 1) as usize..],
-        get_next_era_validators,
-    );
-
-    wait_for_kick_out_event(&root_connection)?;
-
-    // Check current validators.
-    check_validators(
-        &root_connection,
-        &start_era_validators.reserved,
-        &start_era_validators.non_reserved
-            [(VALIDATOR_TO_DISABLE_NON_RESERVED_INDEX + 1) as usize..],
-        get_current_era_validators,
-    );
-    Ok(())
-}
-
-fn setup_test(config: &Config) -> anyhow::Result<(RootConnection, Vec<AccountId>, Vec<AccountId>)> {
-    let root_connection = config.create_root_connection();
-
-    let validator_keys = get_test_validators(config);
-    let reserved_validators = account_ids_from_keys(&validator_keys.reserved);
-    let non_reserved_validators = account_ids_from_keys(&validator_keys.non_reserved);
-
-    let seats = CommitteeSeats {
-        reserved_seats: 2,
-        non_reserved_seats: 2,
-    };
-
-    change_validators(
-        &root_connection,
-        Some(reserved_validators.clone()),
-        Some(non_reserved_validators.clone()),
-        Some(seats),
-        XtStatus::InBlock,
-    );
-
-    wait_for_full_era_completion(&root_connection)?;
-
-    Ok((
-        root_connection,
-        reserved_validators,
-        non_reserved_validators,
-    ))
-}
-
-fn check_validators<C: AnyConnection>(
-    connection: &C,
-    expected_reserved: &[AccountId],
-    expected_non_reserved: &[AccountId],
-    actual_validators_source: fn(&C) -> EraValidators<AccountId>,
-) -> EraValidators<AccountId> {
-    let era_validators = actual_validators_source(connection);
-
-    assert_eq!(era_validators.reserved, expected_reserved);
-    assert_eq!(era_validators.non_reserved, expected_non_reserved);
-
-    era_validators
-}
-
-fn check_committee_kick_out_config(
-    connection: &RootConnection,
-    expected_minimal_expected_performance: Perbill,
-    expected_session_count_threshold: SessionCount,
-    expected_clean_session_counter_delay: SessionCount,
-) -> CommitteeKickOutConfig {
-    let committee_kick_out_config = get_committee_kick_out_config(connection);
-
-    assert_eq!(
-        committee_kick_out_config.minimal_expected_performance,
-        expected_minimal_expected_performance
-    );
-    assert_eq!(
-        committee_kick_out_config.underperformed_session_count_threshold,
-        expected_session_count_threshold
-    );
-    assert_eq!(
-        committee_kick_out_config.clean_session_counter_delay,
-        expected_clean_session_counter_delay
-    );
-
-    committee_kick_out_config
-}
-
-fn check_underperformed_validator_session_count<C: AnyConnection>(
-    connection: &C,
-    validator: &AccountId,
-    expected_session_count: SessionCount,
-) -> SessionCount {
-    let underperformed_validator_session_count =
-        get_underperformed_validator_session_count(connection, validator);
-
-    assert_eq!(
-        underperformed_validator_session_count,
-        expected_session_count
-    );
-
-    underperformed_validator_session_count
-}
-
-fn check_underperformed_validator_reason<C: AnyConnection>(
-    connection: &C,
-    validator: &AccountId,
-    expected_reason: &Option<KickOutReason>,
-) -> Option<KickOutReason> {
-    let validator_kick_out_reason = get_kick_out_reason_for_validator(connection, validator);
-
-    assert_eq!(&validator_kick_out_reason, expected_reason);
-
-    validator_kick_out_reason
-}
-
-#[derive(Debug, Decode, Clone)]
-struct KickOutEvent {
-    kicked_out_validators: Vec<(AccountId, KickOutReason)>,
-}
-
-fn wait_for_kick_out_event<C: AnyConnection>(connection: &C) -> anyhow::Result<KickOutEvent> {
-    let event = wait_for_event(
-        connection,
-        ("Elections", "KickOutValidators"),
-        |e: KickOutEvent| {
-            info!(
-                "Received KickOutValidators event: {:?}",
-                e.kicked_out_validators
-            );
-            true
-        },
-    )?;
-
-    Ok(event)
 }
